refactor(ui): group collapsible component files
