+++ packages/ui/resources/js/components/collapsible/collapsible.browser.test.tsx
