+++ packages/ui/resources/js/components/collapsible/collapsible.test.tsx
